Change population in TreeLife::Block from quint64 to int.
This is really a small change which matters.
The memory footprint reduced by ~10%, and runStep speed improved by ~10%.

+++ src/TreeLife.cpp
@@ -42,7 +42,7 @@ public:
 	static const size_t SIZE = 1 << DEPTH;
 
 	int flag;
-	quint64 population;
+	int population;
 
 	inline void clear()
 	{
